+++ src/plugins/networkFirst.ts
@@ -13,11 +13,13 @@ export function networkFirst(config: NetworkFirstConfig): Plugin {
         order,
         name: 'networkFirst',
 
-        fetch: async (event) => {
+        fetch: async (event, context) => {
             const cache = await caches.open(cacheName);
 
             try {
-                const response = await fetch(event.request);
+                const headers = new Headers(event.request.headers);
+                headers.set(context.passthroughHeader, '1');
+                const response = await fetch(new Request(event.request, { headers }));
 
                 if (response.ok) {
                     await cache.put(event.request, response.clone());